tests: move all unit test modules to osh/tests/
Related: https://github.com/openscanhub/openscanhub/issues/281
Closes: https://github.com/openscanhub/openscanhub/pull/300
