+++ osh.spec
@@ -211,7 +211,6 @@ ln -s %{_sysconfdir}/osh/hub/settings_local.py %{buildroot}%{python3_sitelib}/os
 %{zsh_completions_dir}
 %{python3_sitelib}/osh/client
 %{python3_sitelib}/osh-*-py%{python3_version}.egg-info
-%exclude %{python3_sitelib}/osh/client/tests
 
 %files common
 %dir %{_sysconfdir}/osh
