+++ osh/tests/client/test_cmd_build.py
